Fix spurious indentation changes, format mismatch, and blank lines from code review
Agent-Logs-Url: https://github.com/dotnet/runtime/sessions/c6d266c3-2575-4a12-a8e5-4dcbad30990f

+++ src/coreclr/vm/jitinterface.cpp
@@ -12009,8 +12009,8 @@ void CEEJitInfo::recordRelocation(void *       location,
             }
         }
 
-        LOG((LF_JIT, LL_INFO100000, "Encoded a PCREL32 at" FMT_ADDR "to" FMT_ADDR "+%lld,  delta is 0x%04x\n",
-               DBG_ADDR(fixupLocation), DBG_ADDR(target), (long long)addlDelta, (unsigned int)delta));
+        LOG((LF_JIT, LL_INFO100000, "Encoded a PCREL32 at" FMT_ADDR "to" FMT_ADDR "+%d,  delta is 0x%04x\n",
+             DBG_ADDR(fixupLocation), DBG_ADDR(target), addlDelta, delta));
 
         // Write the 32-bits pc-relative delta into location
         *fixupLocationRW = (INT32) delta;
